requirements.txt
flask
flask-cors
langchain-core
langchain-groq
requests
python-dotenv
